Added some robustness and non clobbering fall backs.

+++ start.sh
@@ -4,12 +4,16 @@ cp -R bashrc/ ~/
 cp -R functions/ ~/
 cp -R bin/ ~/
 
-while read line
-do
-    chmod u+x "${HOME}/bin/$line"
-done < <(ls bin/)
+if [[ $OSTYPE == "msys" ]]; then
+    echo 'Win32 Detected. Cant chmod here.'
+else
+    while read line
+    do
+        chmod u+x "${HOME}/bin/$line"
+    done < <(ls bin/)
+fi
 
-cp .bash_path ~/
-cp .bash_path.opt ~/
+[ ! -f ~/.bash_path ] && cp .bash_path ~/
+[ ! -f ~/.bash_path ] && cp .bash_path.opt ~/
 cp .bashrc ~/
 cp .bash_profile ~/
